add node exporter installer provider
